move resize hook to utils
